@@ -10,12 +10,15 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
+        "php": "^7.4 || ^8.0",
         "ext-dom": "*",
         "ext-libxml": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8 || ^5.7"
+        "phpstan/phpstan": "^0.12.50",
+        "phpstan/phpstan-phpunit": "^0.12.16",
+        "phpunit/phpunit": "^9.4.2",
+        "slam/php-cs-fixer-extensions": "^2.0"
     },
     "autoload": {
         "psr-4": {
