@@ -1,7 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 namespace SlamFatturaElettronica\Exception;
 
-final class InvalidXsdStructureComplianceException extends \ErrorException implements ExceptionInterface
+use ErrorException;
+
+final class InvalidXsdStructureComplianceException extends ErrorException implements ExceptionInterface
 {
 }
