@@ -1,8 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace SlamFatturaElettronica;
 
 interface ValidatorInterface
 {
-    public function assertValidXml($xml);
+    public function assertValidXml(string $xml): void;
 }
